refactor(services): split cell/service.rs into submodules
Decompose the 1,365-line CellService file into a directory of focused
submodules: mod.rs (struct + lifecycle), startup.rs (XML + DB hydration),
message_loop.rs (tick scheduler), base_messages.rs (BaseToCellMsg dispatch),
ticks.rs (AoI/reload/movement ticks), npc_ai.rs (fight + leash), tests.rs.

Public API unchanged. No behavior changes.
