+++ crates/services/src/cell/service/mod.rs
@@ -1,0 +1,96 @@
+//! CellService — spatial entity simulation service.
+//!
+//! In the original C++ architecture, this was the `CellApp` process that:
+//! - Managed game spaces (world zones/instances)
+//! - Simulated cell entity halves (spatial state, movement, AoI)
+//! - Processed entity interactions within spatial proximity
+//! - Ran the game tick loop for entity updates
+
+use std::net::SocketAddr;
+use std::sync::Arc;
+
+use sqlx::PgPool;
+use tokio::sync::mpsc;
+
+use cimmeria_common::ServerConfig;
+
+use super::messages::{BaseToCellMsg, CellToBaseMsg};
+
+mod base_messages;
+mod message_loop;
+mod npc_ai;
+mod startup;
+mod ticks;
+
+#[cfg(test)]
+mod tests;
+
+/// CellApp service managing spatial entity simulation.
+pub struct CellService {
+    /// Address the cell service binds to for BaseApp communication.
+    pub listener_addr: SocketAddr,
+
+    /// Whether the service is currently running.
+    pub is_running: bool,
+
+    /// Receiver for messages from BaseApp (set by orchestrator before start).
+    pub(crate) base_to_cell_rx: Option<mpsc::Receiver<BaseToCellMsg>>,
+
+    /// Sender for messages to BaseApp (set by orchestrator before start).
+    pub(crate) cell_to_base_tx: Option<mpsc::Sender<CellToBaseMsg>>,
+
+    /// Path to the entities directory for loading space XML files.
+    pub(crate) entities_dir: String,
+
+    /// Database pool for content engine loading (set by orchestrator).
+    pub(crate) db_pool: Option<Arc<PgPool>>,
+}
+
+impl CellService {
+    /// Create a new cell service from server configuration.
+    pub fn new(config: &ServerConfig) -> Self {
+        let listener_addr = format!("{}:{}", config.cell_host, config.cell_port)
+            .parse()
+            .unwrap_or_else(|_| {
+                SocketAddr::from(([127, 0, 0, 1], config.cell_port))
+            });
+
+        Self {
+            listener_addr,
+            is_running: false,
+            base_to_cell_rx: None,
+            cell_to_base_tx: None,
+            entities_dir: "entities".to_string(),
+            db_pool: None,
+        }
+    }
+
+    /// Set the database pool for content engine loading.
+    pub fn set_db_pool(&mut self, pool: Arc<PgPool>) {
+        self.db_pool = Some(pool);
+    }
+
+    /// Wire in the Base<->Cell channels. Called by the orchestrator before `start()`.
+    pub fn set_channels(
+        &mut self,
+        rx: mpsc::Receiver<BaseToCellMsg>,
+        tx: mpsc::Sender<CellToBaseMsg>,
+    ) {
+        self.base_to_cell_rx = Some(rx);
+        self.cell_to_base_tx = Some(tx);
+    }
+
+    /// Get a clone of the CellToBase sender (for minigame result routing).
+    pub fn cell_to_base_tx(&self) -> Option<mpsc::Sender<CellToBaseMsg>> {
+        self.cell_to_base_tx.clone()
+    }
+
+    /// Stop the cell service gracefully.
+    pub async fn stop(&mut self) {
+        tracing::info!("Stopping cell service");
+        self.base_to_cell_rx = None;
+        self.cell_to_base_tx = None;
+        self.is_running = false;
+        tracing::trace!("Cell service stopped");
+    }
+}
