+++ crates/services/src/cell/service/npc_ai.rs
@@ -1,0 +1,230 @@
+//! NPC AI tick — fight (threat-target attacks + leashing), and leash recovery.
+//!
+//! Runs every 2 seconds (every 20th AoI tick).
+
+use tokio::sync::mpsc;
+
+use super::super::messages::CellToBaseMsg;
+use super::super::space_manager::SpaceManager;
+
+/// NPC AI tick — for each NPC in Fighting state: find top threat target,
+/// check leash distance, attack with default ability. For NPCs in Leashing
+/// state: reset to Idle when close enough to spawn point, restore health.
+pub(super) async fn npc_ai_tick(
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) {
+    use cimmeria_entity::cell_entity::AiState;
+
+    // Snapshot NPC IDs and their AI state so we don't hold a borrow on space_mgr
+    // while calling handle_use_ability (which needs &mut SpaceManager).
+    let npc_snapshot: Vec<(u32, AiState)> = space_mgr.all_npc_entity_ids()
+        .iter()
+        .filter_map(|&eid| {
+            space_mgr.get_entity(eid).map(|e| (eid, e.ai_state))
+        })
+        .filter(|(_, state)| *state == AiState::Fighting || *state == AiState::Leashing)
+        .collect();
+
+    for (npc_id, ai_state) in npc_snapshot {
+        match ai_state {
+            AiState::Fighting => {
+                npc_ai_fight(npc_id, tx, space_mgr).await;
+            }
+            AiState::Leashing => {
+                npc_ai_leash(npc_id, tx, space_mgr).await;
+            }
+            _ => {}
+        }
+    }
+}
+
+/// NPC fighting behavior: attack top-threat target or leash if too far from spawn.
+async fn npc_ai_fight(
+    npc_id: u32,
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) {
+    use cimmeria_entity::cell_entity::AiState;
+    use super::super::combat;
+
+    // Read NPC state (immutable borrow)
+    let (top_target, spawn_pos, npc_pos) = {
+        let npc = match space_mgr.get_entity(npc_id) {
+            Some(e) => e,
+            None => return,
+        };
+
+        // Find highest-threat target
+        let top = npc.threat_list.iter()
+            .max_by(|a, b| a.1.partial_cmp(b.1).unwrap_or(std::cmp::Ordering::Equal))
+            .map(|(&eid, _)| eid);
+
+        (top, npc.spawn_position, npc.position)
+    };
+
+    let target_id = match top_target {
+        Some(tid) => tid,
+        None => {
+            // No threat targets left — reset to idle
+            if let Some(npc) = space_mgr.get_entity_mut(npc_id) {
+                npc.ai_state = AiState::Idle;
+                npc.threat_list.clear();
+                tracing::debug!(npc_id, "NPC AI: no threat targets, resetting to Idle");
+            }
+            return;
+        }
+    };
+
+    // Check if target still exists and is alive
+    let target_pos = match space_mgr.get_entity(target_id) {
+        Some(t) => {
+            // Don't attack dead targets
+            let is_dead = t.stats.get(cimmeria_entity::stats::HEALTH)
+                .map_or(true, |s| s.cur <= 0);
+            if is_dead {
+                if let Some(npc) = space_mgr.get_entity_mut(npc_id) {
+                    npc.threat_list.remove(&target_id);
+                    tracing::debug!(npc_id, target = target_id, "NPC AI: target is dead, removing from threat");
+                }
+                return;
+            }
+            t.position
+        }
+        None => {
+            // Target gone (disconnected), remove from threat and re-evaluate
+            if let Some(npc) = space_mgr.get_entity_mut(npc_id) {
+                npc.threat_list.remove(&target_id);
+            }
+            return;
+        }
+    };
+
+    // Leash check: if target is too far from NPC's spawn point, disengage
+    if let Some(spawn) = spawn_pos {
+        let dist_to_spawn = spawn.distance_to(&target_pos);
+        if dist_to_spawn > combat::LEASH_DISTANCE {
+            if let Some(npc) = space_mgr.get_entity_mut(npc_id) {
+                npc.ai_state = AiState::Leashing;
+                npc.threat_list.clear();
+                tracing::info!(
+                    npc_id, target = target_id,
+                    distance = dist_to_spawn,
+                    "NPC AI: target too far from spawn, leashing"
+                );
+            }
+            return;
+        }
+    }
+
+    // Range check: don't attack until target is within weapon range
+    let dist_to_target = npc_pos.distance_to(&target_pos);
+    let in_range = dist_to_target <= combat::NPC_ATTACK_RANGE;
+    let has_los = space_mgr.has_line_of_sight(npc_id, target_id);
+
+    if !in_range {
+        // Out of range — pathfind toward target, but only recalculate if:
+        // 1. NPC has no active path, OR
+        // 2. Target has moved significantly from the path's destination
+        let needs_repath = {
+            let npc = space_mgr.get_entity(npc_id);
+            match npc {
+                Some(e) if !e.nav_path.is_empty() => {
+                    // Check if target moved far from the last waypoint
+                    let last_wp = e.nav_path[e.nav_path.len() - 1];
+                    last_wp.distance_to(&target_pos) > 5.0
+                }
+                _ => true, // No path — need one
+            }
+        };
+
+        if needs_repath {
+            if let Some(path) = space_mgr.find_path(npc_id, &npc_pos, &target_pos) {
+                if path.len() > 1 {
+                    let waypoints: Vec<_> = path.into_iter().skip(1).collect();
+                    if let Some(npc) = space_mgr.get_entity_mut(npc_id) {
+                        npc.nav_path = waypoints;
+                    }
+                    tracing::debug!(
+                        npc_id, target = target_id,
+                        in_range, has_los,
+                        "NPC AI: pathfinding toward target"
+                    );
+                }
+            } else {
+                tracing::debug!(
+                    npc_id, target = target_id,
+                    "NPC AI: no path to target"
+                );
+            }
+        }
+        return;
+    }
+
+    // In range — stop moving and attack (LoS check skipped to prevent jittering;
+    // the ability's range check handles the actual validation)
+    if let Some(npc) = space_mgr.get_entity_mut(npc_id) {
+        npc.nav_path.clear();
+    }
+
+    // Attack the target with the default ability
+    tracing::debug!(npc_id, target = target_id, distance = dist_to_target, "NPC AI: attacking top threat target");
+    super::super::abilities::handle_use_ability(
+        npc_id,
+        combat::NPC_DEFAULT_ABILITY,
+        target_id as i32,
+        tx,
+        space_mgr,
+    ).await;
+}
+
+/// NPC leashing behavior: reset to Idle and restore health.
+///
+/// In a full implementation this would pathfind the NPC back to spawn.
+/// For now we snap back instantly and restore health.
+async fn npc_ai_leash(
+    npc_id: u32,
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) {
+    use cimmeria_entity::cell_entity::AiState;
+
+    let (stat_update, state_field) = {
+        let npc = match space_mgr.get_entity_mut(npc_id) {
+            Some(e) => e,
+            None => return,
+        };
+
+        // Snap back to spawn position
+        if let Some(spawn_pos) = npc.spawn_position {
+            npc.position = spawn_pos;
+        }
+
+        // Restore health to max
+        if let Some(health) = npc.stats.get_mut(cimmeria_entity::stats::HEALTH) {
+            health.set_current(health.max);
+        }
+
+        // Clear dead state flag
+        npc.ai_state = AiState::Idle;
+        npc.threat_list.clear();
+        npc.abilities.clear_all_cooldowns();
+
+        // Clear dead flag from the NPC's actual state_field
+        super::super::combat::clear_dead_state(&mut npc.state_field);
+
+        tracing::info!(npc_id, "NPC AI: leash complete, reset to Idle with full health");
+
+        // Collect data before dropping the mutable borrow
+        let stat_update = npc.stats.serialize_dirty();
+        npc.stats.clear_dirty();
+        let state_field = npc.state_field;
+        (stat_update, state_field)
+    };
+
+    super::super::abilities::send_entity_method(npc_id, 20, stat_update, tx, space_mgr).await;
+
+    let mut state_args = Vec::with_capacity(4);
+    state_args.extend_from_slice(&state_field.to_le_bytes());
+    super::super::abilities::send_entity_method(npc_id, 19, state_args, tx, space_mgr).await;
+}
