+++ crates/services/src/cell/service/ticks.rs
@@ -1,0 +1,243 @@
+//! Per-frame tick handlers: AoI propagation, reload-completion promotion,
+//! and NPC movement along nav paths.
+
+use tokio::sync::mpsc;
+
+use super::super::messages::CellToBaseMsg;
+use super::super::space_manager::SpaceManager;
+
+/// Run one tick of AoI processing across all spaces.
+pub(super) async fn run_aoi_tick(
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) {
+    let events = space_mgr.compute_aoi_changes();
+    for event in events {
+        if tx.send(event).await.is_err() {
+            tracing::warn!("Failed to send AoI event to BaseApp (channel closed)");
+            return;
+        }
+    }
+}
+
+/// Promote any reload whose warmup deadline has elapsed: refill the active
+/// bandolier slot's magazine, clear `reload_complete_at`, send `onStatUpdate`
+/// for the AmmoSlot{N} stat to the player, and queue a `BandolierAmmoUpdate`
+/// to base for persistence.
+///
+/// Stage C: this is the sole refill path. The fire-path eager-promotion has
+/// been removed; `handle_use_ability` reads ammo through `entity.active_ammo()`
+/// and the bandolier UI updates on every fire via the AmmoSlot{N} stat.
+pub(super) async fn reload_completion_tick(
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) {
+    let now = std::time::Instant::now();
+
+    // Snapshot ready-to-promote player IDs first to avoid holding a borrow on
+    // `space_mgr` across the `send_entity_method` await below.
+    let ready: Vec<u32> = space_mgr.all_player_entity_ids()
+        .into_iter()
+        .filter(|&eid| {
+            space_mgr.get_entity(eid)
+                .and_then(|e| e.reload_complete_at)
+                .map_or(false, |t| now >= t)
+        })
+        .collect();
+
+    for entity_id in ready {
+        // Phase 1: mutate entity state, capture stat-update payload + the
+        // BandolierAmmoUpdate fields we need to send afterwards. Drop the
+        // mutable borrow before any `.await`.
+        let (stat_payload, persist) = {
+            let entity = match space_mgr.get_entity_mut(entity_id) {
+                Some(e) => e,
+                None => continue,
+            };
+
+            // Refill the slot that *started* the reload, not whatever slot is
+            // currently active. Without pinning, a mid-reload weapon swap
+            // would mis-attribute the refill to the new weapon.
+            let slot_id = match entity.reload_slot_id {
+                Some(s) => s,
+                None => {
+                    // Defensive: shouldn't happen — reload_complete_at is only
+                    // set together with reload_slot_id. Clear and move on.
+                    entity.reload_complete_at = None;
+                    tracing::warn!(entity_id, "reload tick: deadline set without slot_id, clearing");
+                    continue;
+                }
+            };
+
+            // Look up the clip size for the pinned slot. If the slot is
+            // empty (item removed mid-reload), clear the deadline and skip
+            // the wire send rather than refilling nothing.
+            let clip_size = entity.bandolier_items.get(&slot_id).map(|i| i.clip_size);
+            let new_ammo = match clip_size {
+                Some(cs) => entity.set_slot_ammo(slot_id, cs),
+                None => None,
+            };
+            entity.reload_complete_at = None;
+            entity.reload_slot_id = None;
+
+            if new_ammo.is_none() {
+                tracing::debug!(entity_id, slot_id, "reload tick: pinned slot empty, no refill");
+                continue;
+            }
+
+            // The slot was marked dirty by `set_slot_ammo`; persistence drains
+            // it via the BandolierAmmoUpdate below.
+            entity.bandolier_ammo_dirty.remove(&slot_id);
+
+            let payload = entity.stats.serialize_dirty();
+            entity.stats.clear_dirty();
+
+            let (item_id, cur_ammo, cur_ammo_type) = entity.bandolier_items.get(&slot_id)
+                .map_or((0, 0, 0), |i| (i.item_id, i.current_ammo, i.cur_ammo_type));
+            let persist = entity.player_id.map(|pid| (pid, slot_id, item_id, cur_ammo, cur_ammo_type));
+
+            (payload, persist)
+        };
+
+        // Phase 2: send onStatUpdate (method 20) — payload may be empty if
+        // refill was a no-op (e.g. magazine was already at clip_size when the
+        // deadline elapsed because of a concurrent path).
+        if !stat_payload.is_empty() {
+            super::super::abilities::send_entity_method(entity_id, 20, stat_payload, tx, space_mgr).await;
+        }
+
+        // Phase 3: persistence. CellToBaseMsg::BandolierAmmoUpdate is consumed
+        // by base's existing handler that writes `sgw_inventory.ammo`.
+        if let Some((player_id, slot_id, expected_item_id, current_ammo, cur_ammo_type)) = persist {
+            let _ = tx.send(CellToBaseMsg::BandolierAmmoUpdate {
+                player_id,
+                slot_id,
+                expected_item_id,
+                current_ammo,
+                cur_ammo_type,
+            }).await;
+        }
+    }
+}
+
+/// NPC movement along nav paths — runs every AoI tick (100ms) for smooth pathing.
+///
+/// For each NPC with a non-empty `nav_path`, move it toward the next waypoint
+/// by `move_speed` units. When it reaches (or overshoots) a waypoint, consume
+/// it and continue to the next. Position updates propagate to witnesses via
+/// the AoI tick's `EntityMoved` messages.
+pub(super) fn npc_movement_tick(space_mgr: &mut SpaceManager) {
+    // Collect NPCs that have active paths
+    let moving_npcs: Vec<u32> = space_mgr.all_npc_entity_ids()
+        .iter()
+        .filter(|&&eid| {
+            space_mgr.get_entity(eid)
+                .map_or(false, |e| !e.nav_path.is_empty())
+        })
+        .copied()
+        .collect();
+
+    for npc_id in moving_npcs {
+        // Read the next waypoint, move_speed, and remaining path length
+        let (next_wp, move_speed, cur_pos, path_len) = {
+            let npc = match space_mgr.get_entity(npc_id) {
+                Some(e) if !e.nav_path.is_empty() => e,
+                _ => continue,
+            };
+            (npc.nav_path[0], npc.move_speed, npc.position, npc.nav_path.len())
+        };
+
+        let dx = next_wp.x - cur_pos.x;
+        let dy = next_wp.y - cur_pos.y;
+        let dz = next_wp.z - cur_pos.z;
+        let dist = (dx * dx + dy * dy + dz * dz).sqrt();
+
+        // Speed in world units per second (tick is 100ms = 0.1s)
+        let speed_per_sec = move_speed * 10.0;
+
+        if dist <= move_speed {
+            // Reached (or overshot) the waypoint — snap to it and consume
+            // Waypoint Y comes from Detour's findStraightPath (already on navmesh surface)
+            let snap_y = next_wp.y;
+
+            // Peek at the NEXT waypoint (index 1) to compute velocity toward it
+            let next_next_wp = if path_len > 1 {
+                space_mgr.get_entity(npc_id).map(|e| e.nav_path[1])
+            } else {
+                None
+            };
+
+            let (velocity, yaw) = if let Some(nn) = next_next_wp {
+                // Still more waypoints — compute velocity toward the next one
+                let ndx = nn.x - next_wp.x;
+                let ndz = nn.z - next_wp.z;
+                let ndy = nn.y - next_wp.y;
+                let nd = (ndx * ndx + ndy * ndy + ndz * ndz).sqrt();
+                if nd > 0.001 {
+                    (
+                        [ndx / nd * speed_per_sec, ndy / nd * speed_per_sec, ndz / nd * speed_per_sec],
+                        ndx.atan2(ndz),
+                    )
+                } else {
+                    ([0.0; 3], 0.0)
+                }
+            } else {
+                // Last waypoint — stopping, keep current facing
+                ([0.0; 3], dx.atan2(dz))
+            };
+
+            space_mgr.update_entity_position(
+                npc_id,
+                [next_wp.x, snap_y, next_wp.z],
+                [0, 0, 0],
+                velocity,
+            );
+            if let Some(npc) = space_mgr.get_entity_mut(npc_id) {
+                npc.nav_path.remove(0);
+                npc.direction = cimmeria_common::Vector3::new(0.0, yaw, 0.0);
+            }
+        } else {
+            // Move toward waypoint by move_speed units
+            let t = move_speed / dist;
+            let new_x = cur_pos.x + dx * t;
+            let new_z = cur_pos.z + dz * t;
+
+            // Linearly interpolate Y between current position and waypoint.
+            // Waypoints from Detour's findStraightPath are on the navmesh surface,
+            // so linear interpolation between them stays close to the floor.
+            let new_y = cur_pos.y + dy * t;
+
+            // Face the direction of movement (yaw = atan2(dx, dz) in radians)
+            // Direction is [pitch, yaw, roll] — only yaw matters for facing
+            let yaw = dx.atan2(dz);
+
+            // Velocity = direction * speed_per_sec
+            let velocity = [
+                dx / dist * speed_per_sec,
+                dy / dist * speed_per_sec,
+                dz / dist * speed_per_sec,
+            ];
+
+            if (npc_id % 10000) < 5 { // log a few NPCs
+                tracing::debug!(
+                    npc_id,
+                    cur = format_args!("({:.1},{:.1},{:.1})", cur_pos.x, cur_pos.y, cur_pos.z),
+                    new = format_args!("({:.1},{:.1},{:.1})", new_x, new_y, new_z),
+                    wp = format_args!("({:.1},{:.1},{:.1})", next_wp.x, next_wp.y, next_wp.z),
+                    "NPC movement step"
+                );
+            }
+
+            space_mgr.update_entity_position(
+                npc_id,
+                [new_x, new_y, new_z],
+                [0, 0, 0],
+                velocity,
+            );
+            // Set yaw directly as radians (pack_angle reads direction.y)
+            if let Some(npc) = space_mgr.get_entity_mut(npc_id) {
+                npc.direction = cimmeria_common::Vector3::new(0.0, yaw, 0.0);
+            }
+        }
+    }
+}
